relacionamento usuário -> notificação

+++ gas-out-app/src/main/java/br/com/gasoutapp/service/NotificationService.java
@@ -1,5 +1,6 @@
 package br.com.gasoutapp.service;
 
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -7,15 +8,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.gasoutapp.domain.Notification;
+import br.com.gasoutapp.domain.User;
 import br.com.gasoutapp.dto.NotificationDTO;
 import br.com.gasoutapp.exception.NotificationNotFoundException;
+import br.com.gasoutapp.exception.UserNotFoundException;
 import br.com.gasoutapp.repository.NotificationRepository;
+import br.com.gasoutapp.repository.UserRepository;
 
 @Service
 public class NotificationService {
 
 	@Autowired
 	private NotificationRepository notificationRepository;
+	
+	@Autowired
+	private UserRepository userRepository;
 	
 	public List<Notification> getAllNotifications() {
 		return notificationRepository.findAll();
@@ -28,20 +35,41 @@ public class NotificationService {
 		return notifications;
 	}
 
-	public Notification createNotification(NotificationDTO dto) {
-		
-		List<Notification> notifications = getAllNotifications();
+	public Notification createNotification(NotificationDTO dto) {	
+		List<Notification> newUserNotifications = new ArrayList<>();
+		User newUser = new User();
+		User user = userRepository.findByLogin(dto.getEmail());
+		if(user == null) {
+			throw new UserNotFoundException();
+		} 
+		newUser = user;
+				
+		List<Notification> notifications = notificationRepository.findAllByUser(user);
 		if(notifications.size() >= 10) {
 			deleteAllNotifications(notifications);
+		} else {
+			newUserNotifications = notifications;
 		}
 		
 		Notification newNotification = new Notification();
 		
+		newNotification.setUser(user);
 		newNotification.setTitle(dto.getTitle());
 		newNotification.setMessage(dto.getMessage());
 		newNotification.setDate(new Date());
 		
-		return notificationRepository.save(newNotification);
+		newNotification = notificationRepository.save(newNotification);
+		newUserNotifications.add(newNotification);
+		
+		newUser.setNotifications(newUserNotifications);
+		userRepository.save(newUser);
+		
+		List<Notification> notificationsUserNull = notificationRepository.findAllByUser(null);
+		if(notificationsUserNull.size() > 0) {
+			notificationRepository.deleteAll(notificationsUserNull);			
+		}
+				
+		return newNotification;
 	}
 
 	public void deleteNotification(Long id) {
@@ -54,7 +82,20 @@ public class NotificationService {
 	}
 	
 	public void deleteAllNotifications(List<Notification> notifications) {
-		notificationRepository.deleteAll(notifications);
+		for(Notification notification: notifications) {		
+			notification.setUser(null);
+			notification = notificationRepository.save(notification);
+		}		
+	}
+	
+	public Integer deleteAllNoUsersNotifications(List<Notification> noUserNotifications) {
+		if(noUserNotifications.size() > 0) {
+			for(Notification n: noUserNotifications) {
+				notificationRepository.delete(n);				
+			}
+		}
+		
+		return noUserNotifications.size();
 	}
 	
 	public static<T> void reverseList(List<T> list) {
